Add pre-evaluation lookahead to complement post-evaluation lookahead
Previously, hacks were used that allowed enfixed operators to quote
their left argument if they were SET-WORD! and SET-PATH!, since the
operations were pending and still available through that pending
stack.  These hacks were insufficient to allow operators which would
quote their left-hand side if they were evaluative, since the only
lookahead phase happened *after* the evaluation.

This adds a lookahead step *before* the evaluation to address the
quoting issues.  Since lookahead involves dereferencing a WORD! to
get its value cell to see if it is holding an enfixed function, this
has an associated cost of making a Get_Var() call on the word.  To
mitigate the number of calls to Get_Var(), this enhances the previous
mechanism for reusing the fetch (f->gotten) to allow for the pre
evaluation lookahead to participate as well.

Prefix quoting operators running from the left have precedence over
ones running to the right.  The ramifications of this are:

    >> left-quote: enfix :quote

    >> + left-quote ;-- no contention, left-quotes as is
    == +

    >> quote left-quote ;-- prefix quote will win
    == left-quote

    >> quote + left-quote ;-- paradox, must error
    ** Script error: LEFT-QUOTE can't left quote a forward quoted value